src/api-engine/requirements.txt
amqp==5.2.0
asgiref==3.3.1
async-timeout==4.0.3
attrs==20.3.0
billiard==3.6.4.0
cachetools==4.1.1
celery==5.2.6
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.14
django-allauth==0.41.0
django-cors-headers==3.5.0
django-extensions==2.2.9
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.2
djangorestframework-simplejwt==5.2.2
docker==4.2.0
drf-yasg==1.17.1
flex==6.14.1
google-auth==1.23.0
holdup==1.8.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonpointer==2.0
jsonschema==3.2.0
kombu==5.3.5
kubernetes==11.0.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pathtools==0.1.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygraphviz==1.10
PyJWT==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.4.4
requests==2.31.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.4.4
strict-rfc3339==0.7
swagger-spec-validator==2.5.0
uritemplate==3.0.1
urllib3==1.26.17
uWSGI==2.0.22
validate-email==1.3
vine==5.1.0
watchdog==0.10.2
wcwidth==0.2.13
websocket-client==0.57.0
